Move assert function in _internal.core
