+++ utils/manifest/__init__.py
@@ -1,4 +1,12 @@
 from ._internal import Manifest
 from ._internal import SkipDeclaration, Condition, ManifestData, TestDeclaration
+from ._internal import assert_versions_not_ahead_of_current
 
-__all__ = ["Condition", "Manifest", "ManifestData", "SkipDeclaration", "TestDeclaration"]
+__all__ = [
+    "Condition",
+    "Manifest",
+    "ManifestData",
+    "SkipDeclaration",
+    "TestDeclaration",
+    "assert_versions_not_ahead_of_current",
+]
