+++ utils/manifest/_internal/__init__.py
@@ -1,5 +1,12 @@
 from .const import TestDeclaration
-from .core import Manifest
+from .core import Manifest, assert_versions_not_ahead_of_current
 from .types import SkipDeclaration, Condition, ManifestData
 
-__all__ = ["Condition", "Manifest", "ManifestData", "SkipDeclaration", "TestDeclaration"]
+__all__ = [
+    "Condition",
+    "Manifest",
+    "ManifestData",
+    "SkipDeclaration",
+    "TestDeclaration",
+    "assert_versions_not_ahead_of_current",
+]
